zcash_pool_migration: move the greedy into preparation/, behind plan_preparation
Split the module along the line the strategy seam drew. preparation.rs keeps
the PROBLEM: the plan types, PrepError, the PreparationStrategy trait, the
PlanQuality order, the certificate, the portfolio fold, and the shared entry
validation. Each rule for solving it gets a module of its own under
preparation/, so adding one is adding a file rather than growing this one;
layered_greedy.rs is the first, holding the largest-first layered greedy and
the tests that exercise it. The strategy prose moves out of the parent's
module docs with it, and gains a section naming what that rule cannot reach.

plan_preparation keeps its signature, its contract and its path, but is now
the portfolio rather than one rule's entry point: it validates the instance,
runs every strategy in STRATEGIES, and returns the best plan any of them
produced. Which rule wins becomes an implementation detail, so a strategy
added to that list reaches every caller, the engine included, without a
caller changing. The crate ships one strategy today, so the plans are
unchanged, which the tests assert both ways: the portfolio equals the greedy,
and the public entry point equals the portfolio.

Validating that the note totals are representable amounts is a property of
the INPUTS rather than of any rule, so it moves to a shared validate_instance
that plan_preparation runs once before dispatching. A strategy invoked
directly through the trait cannot rely on that, so each still calls it at
entry; the planners' internal arithmetic partitions those totals and would
otherwise be assuming what it has not checked.
